# Timed Difficulty Attributes calculation optimization