--- a/osu.Game.Rulesets.Osu/Difficulty/OsuDifficultyCalculator.cs
+++ b/osu.Game.Rulesets.Osu/Difficulty/OsuDifficultyCalculator.cs
@@ -82,13 +82,25 @@ namespace osu.Game.Rulesets.Osu.Difficulty
             double preempt = IBeatmapDifficultyInfo.DifficultyRange(beatmap.Difficulty.ApproachRate, 1800, 1200, 450) / clockRate;
             double drainRate = beatmap.Difficulty.DrainRate;
 
-            ProgressiveCalculationBeatmap pcBeatmap = beatmap as ProgressiveCalculationBeatmap;
+            int maxCombo;
+            int hitCirclesCount, sliderCount, spinnerCount;
 
-            int maxCombo = pcBeatmap != null ? pcBeatmap.GetMaxCombo() : beatmap.GetMaxCombo();
+            if (beatmap is ProgressiveCalculationBeatmap pcBeatmap)
+            {
+                maxCombo = pcBeatmap.GetMaxCombo();
 
-            int hitCirclesCount = pcBeatmap != null ? pcBeatmap.GetHitObjectCountOf(typeof(HitCircle)) : beatmap.HitObjects.Count(h => h is HitCircle);
-            int sliderCount = pcBeatmap != null ? pcBeatmap.GetHitObjectCountOf(typeof(Slider)) : beatmap.HitObjects.Count(h => h is Slider);
-            int spinnerCount = pcBeatmap != null ? pcBeatmap.GetHitObjectCountOf(typeof(Spinner)) : beatmap.HitObjects.Count(h => h is Spinner);
+                hitCirclesCount = pcBeatmap.GetHitObjectCountOf(typeof(HitCircle));
+                sliderCount = pcBeatmap.GetHitObjectCountOf(typeof(Slider));
+                spinnerCount = pcBeatmap.GetHitObjectCountOf(typeof(Spinner));
+            }
+            else
+            {
+                maxCombo = beatmap.GetMaxCombo();
+
+                hitCirclesCount = beatmap.HitObjects.Count(h => h is HitCircle);
+                sliderCount = beatmap.HitObjects.Count(h => h is Slider);
+                spinnerCount = beatmap.HitObjects.Count(h => h is Spinner);
+            }
 
             HitWindows hitWindows = new OsuHitWindows();
             hitWindows.SetDifficulty(beatmap.Difficulty.OverallDifficulty);