Timed Difficulty Attributes used for pp counter, and it's very slow on long maps because of O(n^2 * logn) time complexity

Test on full Because Maybe mapset (https://osu.ppy.sh/beatmapsets/773014#osu/2151273). Out of 152 seconds of main function call:
- 68 seconds on OsuStrainSkill.DifficultyValue()
- 52 seconds on IBeatmap.GetMaxCombo()
- 29 seconds on IBeatmap.HitObjects.Count(h => h is HitObject)
- 3 seconds on Speed.RelevantNoteCount()

Out of 68 seconds of OsuStrainSkill.DifficultyValue():
- 48 seconds on initial sort
- 17 seconds on second sort (after diffspike nerf)

After the optimization - out of 13 seconds (-91%) of main function call:
- 10 seconds (-85%) on OsuStrainSkill.DifficultyValue()
- 0.06 seconds (-99.9%) on IBeatmap.GetMaxCombo()
- 0.07 seconds (-99.8%) ( on IBeatmap.HitObjects.Count(h => h is HitObject)
- 3 seconds (not touched) on Speed.RelevantNoteCount()

Out of 10 seconds of OsuStrainSkill.DifficultyValue():
- 6 seconds (-87.5%) on initial sort, 5 seconds of which is creating a copy, and only 1 second is actual sorting. Considering this - speedup on actual sorting is 43x (-97.7%)
- 0.5 seconds (-97%) on the second sort (after diffspike nerf)

Optimization consists of two parts:
First part makes Max Combo and Objects Counts to be calculated progressively in the `ProgressiveBeatmap` beatmap class instead of from zero each time, reducing time complexity from O(n^2) to O(n).

Second part is all about `.DifficultyValue()`:
After the call of `DifficultyValue()` the sorted strains are saved in the Skill class. On the next call of `DifficultyValue` one of 4 things can happen:
1) There are no saved strains -> calculate normally
2) There are saved strains, but new ones were added since last save -> insert each strain in the sorted array
3) There are no new strains since save, but the last strain was changed -> replace old last strain with new one
4) No new strains and last strain is unchanged -> just use the saved strains
Those steps are reducing sort time complexity from O(n^2 * logn) to O(nlogn). Copying time complexity remains O(n^2), making it the slowest part of the calculation now.

The second sort after diffspike nerf is also changed. Unnecessary copy was removed and sorting is done via insertions now, decreasing O(n^2) for copy and O(n^2 * logn) for sort to just O(nlogn) for insertion.

Before:

https://github.com/user-attachments/assets/eb18f836-86e0-4e8f-a8eb-ef29897d6b27

After:

https://github.com/user-attachments/assets/98521e8d-f81f-4832-9bdf-95075f3a76c8

The other gamemodes aren't touched directly, but as they all use `StrainSkill` - they will be faster as well. Taiko needs additional work, as it's still have `TaikoDifficultyCalculator.combinedDifficultyValue` with O(n^2 * logn) time complexity.