@Joehuu 2024-08-17
We don't have that label. https://github.com/ppy/osu/labels/type%3Aperformance is talking about the game performance. Will just unlabel.