@Givikap120 2024-08-17
> We don't have that label. [ type:performance](https://github.com/ppy/osu/labels/type%3Aperformance) is talking about the game performance. Will just unlabel.

This is about game performance, no?
Reducing run time of the function