@Givikap120 2024-08-19
Also, it looks like Extension Method `GetMaxCombo` is only correct for STD and maybe for Taiko. For now I just copied it's code into the respective function, but it looks like oversight for me.