@Givikap120 2024-08-20
> > Also, it looks like Extension Method `GetMaxCombo` is only correct for STD and maybe for Taiko.
> 
> Huh? Care to elaborate? It looks correct to me.

Just checked different maps: alongside Osu and Taiko - Catch is correct too. However - Mania is definitely incorrect because of Long Notes giving combo without judgements.