@Givikap120 2024-08-20
> > Mania is definitely incorrect because of Long Notes giving combo without judgements
> 
> mania is correct for lazer mechanics. "long notes giving combo without judgements" is a stable thing we didn't reimplement.

Difficulty calculator needs to account for legacy scores (with CL?). Or am I incorrect?