@Givikap120 2024-08-20
> It already does though?
> 
> https://github.com/ppy/osu/blob/ebff2ebf012b4cc6999bbedd2e99cdccef2060e4/osu.Game.Rulesets.Mania/Difficulty/ManiaDifficultyCalculator.cs#L62-L68

Every ruleset-specific (except STD) calculator using it's own Max Combo function instead of extension function. And in case of mania - then it's only correct for Legacy scores and incorrect for new Lazer mania combo system.