@Givikap120 2024-08-20
> That may well be true but that entire conversation is out of scope of this pull which is supposed to be an "optimisation". I'd advise to move it elsewhere.

okay I agree
then I just waiting on review